requirements.txt
python-telegram-bot
langchain-gigachat
langchain
chromadb
langchain-chroma
aiologger
pydantic
pydantic-settings
dotenv